requirements.txt
absl-py==2.1.0
astunparse==1.6.3
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
filelock==3.12.2
Flask==2.2.5
flatbuffers==24.3.25
gast==0.4.0
gdown==4.7.3
google-auth==2.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.3
h5py==3.8.0
idna==3.10
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.4
keras==2.11.0
libclang==18.1.1
Markdown==3.4.4
MarkupSafe==2.1.5
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.0
pandas==1.3.5
Pillow==9.5.0
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
soupsieve==2.4.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.17.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tqdm==4.66.5
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
wrapt==1.16.0
zipp==3.15.0
